Consultant

+++ lib/Layouts/SideBar_Layouts/Setting.dart
@@ -10,11 +10,10 @@ class Setting extends StatefulWidget {
   const Setting({Key? key}) : super(key: key);
 
   @override
-  State<Setting> createState() => _SettingsState();
+  State<Setting> createState() => _SettingState();
 }
 
-class _SettingsState extends State<Setting> {
-
+class _SettingState extends State<Setting> {
   void signOut(){
     FirebaseAuth.instance.signOut().onError((error, stackTrace){
       ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(error.toString())));
@@ -28,11 +27,12 @@ class _SettingsState extends State<Setting> {
 
       await FirebaseAuth.instance.currentUser!.delete();
 
-      await documentReference.delete();
-      print('Account deleted successfully.');
+      await documentReference.delete()
+          .then((value) {
+        ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text("Account deleted successfully."),duration: Duration(seconds: 2),));
+      });
     } catch (e) {
-      // Error message
-      print('Error Deleting account: $e');
+      ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(e.toString()),duration: Duration(seconds: 3),));
     }
   }
 
@@ -55,11 +55,11 @@ class _SettingsState extends State<Setting> {
             SizedBox(height: 20,),
 
             Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 20),
+              padding: const EdgeInsets.symmetric(horizontal: 17),
               child: Container(
                 decoration: BoxDecoration(
-                    color: Colors.white,
-                    borderRadius: BorderRadius.circular(15)
+                    color: Colors.white70,
+                    borderRadius: BorderRadius.circular(25)
                 ),
                 child: ListTile(
                   onTap: (){
@@ -75,11 +75,11 @@ class _SettingsState extends State<Setting> {
             SizedBox(height: 15,),
 
             Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 20),
+              padding: const EdgeInsets.symmetric(horizontal: 17),
               child: Container(
                 decoration: BoxDecoration(
-                    color: Colors.white,
-                    borderRadius: BorderRadius.circular(15)
+                    color: Colors.white70,
+                    borderRadius: BorderRadius.circular(25)
                 ),
                 child: ListTile(
                   onTap: (){
@@ -95,11 +95,11 @@ class _SettingsState extends State<Setting> {
             SizedBox(height: 15,),
 
             Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 20),
+              padding: const EdgeInsets.symmetric(horizontal: 17),
               child: Container(
                 decoration: BoxDecoration(
-                    color: Colors.white,
-                    borderRadius: BorderRadius.circular(15)
+                    color: Colors.white70,
+                    borderRadius: BorderRadius.circular(25)
                 ),
                 child: ListTile(
                   onTap: (){
@@ -115,19 +115,39 @@ class _SettingsState extends State<Setting> {
             SizedBox(height: 15,),
 
             Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 20),
+              padding: const EdgeInsets.symmetric(horizontal: 17),
               child: Container(
                 decoration: BoxDecoration(
-                  color: Colors.white,
-                  borderRadius: BorderRadius.circular(15)
+                  color: Colors.white70,
+                  borderRadius: BorderRadius.circular(25)
                 ),
                 child: ListTile(
                   onTap: (){
-                    signOut();
-                    Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => LoginPage(),));},
+                    showDialog(
+                        context: context,
+                        builder: (ctx) => AlertDialog(
+                        title: Text('Log Out',style: TextStyle(fontSize: 30),),
+                        content: Text('Are you sure want to Log Out ?',style: TextStyle(fontSize: 17),),
+                        actions: [
+                          TextButton(
+                            onPressed: () {
+                              Navigator.pop(context);
+                            },
+                            child: Text('NO'),
+                          ),
+                          TextButton(
+                            onPressed: () {
+                                signOut();
+                                Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => LoginPage(),));
+                              },
+                            child: Text('YES'),
+                          ),
+                        ],
+                       )
+                    );
+                  },
                   leading: Icon(Icons.logout, size: 35, color: Colors.black,),
                   title: Text("\t\tLog Out",style: TextStyle(fontWeight: FontWeight.w500,fontSize: 20),),
-
                 ),
               ),
             ),
@@ -135,16 +155,36 @@ class _SettingsState extends State<Setting> {
             SizedBox(height: 15,),
 
             Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 20),
+              padding: const EdgeInsets.symmetric(horizontal: 17),
               child: Container(
                 decoration: BoxDecoration(
-                    color: Colors.white,
-                    borderRadius: BorderRadius.circular(15)
+                    color: Colors.white70,
+                    borderRadius: BorderRadius.circular(25)
                 ),
                 child: ListTile(
                   onTap: (){
-                    deleteAccount(FirebaseAuth.instance.currentUser!.uid);
-                    Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => LoginPage(),));
+                    showDialog(
+                      context: context,
+                      builder: (ctx) => AlertDialog(
+                        title: Text('Delete Account',style: TextStyle(fontSize: 30),),
+                        content: Text('Are you sure want to Delete Account ?',style: TextStyle(fontSize: 17),),
+                        actions: [
+                          TextButton(
+                            onPressed: () {
+                              Navigator.pop(context);
+                            },
+                            child: Text('NO'),
+                          ),
+                          TextButton(
+                            onPressed: () {
+                              deleteAccount(FirebaseAuth.instance.currentUser!.uid);
+                              Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => LoginPage(),));
+                            },
+                            child: Text('YES'),
+                          ),
+                        ],
+                     )
+                    );
                   },
                   leading: Icon(Icons.delete_outline, size: 35, color: Colors.black,),
                   title: Text("\t\tDelete Account",style: TextStyle(fontWeight: FontWeight.w500,fontSize: 20),),
